requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.7.7
amqp==5.2.0
annotated-types==0.6.0
anyio==3.6.1
apache-airflow==2.3.1
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==2.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==20.3.0
Babel==2.10.1
backoff==2.2.1
billiard==4.2.0
blinker==1.4
cachelib==0.7.0
cattrs==1.10.0
celery==5.3.5
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.13.1
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
exceptiongroup==1.1.3
Flask==1.1.2
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Limiter==3.5.0
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flower==2.0.1
frozenlist==1.4.0
google-re2==1.1
googleapis-common-protos==1.61.0
graphviz==0.20
greenlet==1.1.2
grpcio==1.59.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
humanize==4.8.0
idna==3.3
importlib-metadata==6.8.0
importlib-resources==6.1.1
inflection==0.5.1
iso8601==2.1.0
itsdangerous==1.1.0
Jinja2==3.0.3
joblib==1.3.2
jsonschema==4.5.1
jsonschema-specifications==2023.7.1
kombu==5.3.3
lazy-object-proxy==1.7.1
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.0
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python==8.2.0
nltk==3.8.1
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
ordered-set==4.1.0
packaging==21.3
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
praw==7.7.1
prawcore==2.4.0
prison==0.2.1
prometheus-client==0.18.0
prompt-toolkit==3.0.40
protobuf==4.21.12
psutil==5.9.1
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.12.0
PyJWT==1.7.1
pymongo==3.11.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.3.1
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.27.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==12.4.1
rich-argparse==1.4.0
rpds-py==0.12.0
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.4
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
textblob==0.17.1
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.2.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
update-checker==0.18.0
urllib3==1.26.9
vine==5.1.0
wcwidth==0.2.10
websocket-client==1.6.4
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
yarl==1.9.2
zipp==3.17.0
